+++ torch/_dynamo/debug_utils.py
@@ -243,63 +243,6 @@ torch._functorch.config.load_config({repr(torch._functorch.config.save_config())
 TEST_REPLACEABLE_COMMENT = "# REPLACEABLE COMMENT FOR TESTING PURPOSES"
 
 
-INDUCTOR_IMPORT = """
-from torch._inductor.compile_fx import compile_fx_inner
-from torch._dynamo.debug_utils import same_two_models
-"""
-
-
-COMPILER_REPRO_OPTIONS = {
-    "inductor": (INDUCTOR_IMPORT, "compile_fx_inner", "inductor_fails"),
-    "inductor_accuracy": (
-        INDUCTOR_IMPORT,
-        "compile_fx_inner",
-        "inductor_accuracy_fails",
-    ),
-}
-
-
-def inductor_fails(fx_g, args, check_str=None):
-    has_cuda = False
-    for arg in args:
-        if arg.is_cuda:
-            has_cuda = True
-            break
-
-    def sync():
-        if has_cuda:
-            # Ensures that segfaults are surfaced
-            torch.cuda.synchronize()
-
-    from torch._inductor.compile_fx import compile_fx_inner
-
-    try:
-        result = fx_g(*args)
-        assert isinstance(result, (tuple, list))
-        assert not any([isinstance(x, (tuple, list)) for x in result])
-    except Exception:
-        return False
-
-    sync()
-
-    try:
-        compile_mod = compile_fx_inner(fx_g, args)
-        compile_mod(args)
-        sync()
-    except Exception as e:
-        if check_str is not None and check_str not in repr(e):
-            return False
-        print(repr(e))
-        return True
-    return False
-
-
-def inductor_accuracy_fails(fx_g, args, check_str=None):
-    from torch._inductor.compile_fx import compile_fx_inner
-
-    return backend_aot_accuracy_fails(fx_g, args, compile_fx_inner)
-
-
 def get_minifier_repro_path():
     return os.path.join(minifier_dir(), "minifier_launcher.py")
 
@@ -464,8 +407,3 @@ def backend_accuracy_fails(gm, example_inputs, compiler_fn, only_fwd=False):
         return False
 
     return not same_two_models(gm, compiled_gm, example_inputs, only_fwd)
-
-
-# NB: This lives here because inductor_accuracy_fails relies on it; this
-# is only used by after_aot
-backend_aot_accuracy_fails = functools.partial(backend_accuracy_fails, only_fwd=True)
